--- config.json
@@ -1,3 +1,0 @@
-{
-  "NARMS": "http://193.10.30.126/api/cals/"
-}
