+++ router/middleware/config.json
@@ -1,0 +1,3 @@
+{
+  "NARMS": "http://193.10.30.126/api/cals/"
+}
